Rename the json_parse to json. TODO: make it can dump the json from typed pipe as well
